Use luassert match.is_ref to check for same instance in tests
Busted tests 'should call start on the current node on first run'
in branch_node_spec, priority_spec, random_spec, and sequence_spec
were failing because the call:
  assert.stub(node.start).was.called_with(node, 'foo')
would fail due to node being modified between when the stub was
created and the assert. See this issue in luassert:
  https://github.com/Olivine-Labs/luassert/issues/123

The correct course of action was to check that node was a reference
to the same object the stub had been created on, using luassert
match.is_ref(node). I've made this change for all tests where it
seemed to apply.