Enhance: Add ActivePriority branch node type
An active priority node acts like a priority node, except that it
ALWAYS runs its children from left to right, even when one of its
children previously called 'running()'. This allows it to react
to changes in the environment that make one of its higher-priority
children a better choice than the currently-running, lower-priority
child. When a higher-priority behaviour takes precedence, the
current implementation calls finish() on the previously-running
child to inform it that it has been pre-empted.